// target-react/src/modules/rubros/domain/model/RubrosidomainRubrosidolistRubrosidodetail.ts
// Generated by migration-flow
// Source: migration/plan/config.json
// GeneratedAt: 2026-03-27

export interface RubrosidomainRubrosidolistRubrosidodetail {
  id: number;
  rubroId: string;
  colaboradorId: number;
  empresaId: number;
  tipoComportamiento?: number;
  valor01?: number;
  valor02?: number;
  valor03?: number;
  estado?: string;
  fechaDesde?: string;
  fechaHasta?: string;
  nombreColaborador?: string;
  nombreRubro?: string;
  fechaCreacion?: string;
  fechaActualizacion?: string;
}

export type CreateRubrosidomainRubrosidolistRubrosidodetail = {
  rubroId: string;
  colaboradorId: number;
  empresaId: number;
  tipoComportamiento?: number;
  valor01?: number;
  valor02?: number;
  valor03?: number;
  estado?: string;
  fechaDesde?: string;
  fechaHasta?: string;
};

export type UpdateRubrosidomainRubrosidolistRubrosidodetail = {
  tipoComportamiento?: number;
  valor01?: number;
  valor02?: number;
  valor03?: number;
  estado?: string;
  fechaHasta?: string;
};

export interface RubrosidomainRubrosidolistRubrosidodetailFilter {
  page?: number;
  size?: number;
  rubroId?: string;
  colaboradorId?: number;
  empresaId?: number;
  estado?: string;
  fechaDesde?: string;
  fechaHasta?: string;
}

export interface RubrosidomainRubrosidolistRubrosidodetailPageResult {
  items: RubrosidomainRubrosidolistRubrosidodetail[];
  totalElements: number;
  totalPages: number;
  currentPage: number;
}